backend_drf/requirements.txt
Django>=6.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.9.0
django-filter>=25.2
Pillow>=12.0.0

